Fix remaining SDK v3 build errors
- PptxConverter: remove SpaceBefore spacing (A.SpacePercent type not in SDK v3)
- SlideRenderer: Width.Value is Int32Value (int), not long — cast (long) for EmuToPx
- PresentationModel: replace ColorSchemeIndexValues enum references with BuildDefaultColorMap() helper that sets raw XML attributes — ColorSchemeIndexValues cannot be resolved via using in SDK v3.2.0

+++ PPEditer/Rendering/SlideRenderer.cs
@@ -155,7 +155,7 @@ public static class SlideRenderer
         var c = ResolveColor(solid, slidePart);
         if (!c.HasValue) return;
         rect.Stroke          = new SolidColorBrush(c.Value);
-        rect.StrokeThickness = ln?.Width?.Value is long thick ? EmuToPx(thick) : 1;
+        rect.StrokeThickness = ln?.Width?.Value is int thick ? EmuToPx((long)thick) : 1;
     }
 
     // ── Text body ─────────────────────────────────────────────────────
